twitterbot/requirements.txt
python-decouple==3.8
jinja2
mongoengine
fastapi_jwt_auth
uvicorn
